pyproject.toml
[tool.poetry]
name = "beam"
version = "0.1.0"
description = ""
authors = ["Aidyn"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.0"
Django = "^5.0.3"
psycopg2-binary = "^2.9.9"
pillow = "^10.2.0"
celery = "^5.3.6"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
django-filter = "^24.1"
python-decouple = "^3.8"
rest-framework-simplejwt = "^0.0.2"
djangorestframework-simplejwt = "^5.3.1"
pymemcache = "^4.0.0"
flake8 = "^7.0.0"
black = "^24.3.0"
stripe = "^8.8.0"
django-countries = "^7.5.1"
docker = "^7.0.0"
redis = "^5.0.3"
dj-database-url = "^2.2.0"
django-stubs-ext = "^5.1.0"
django-jazzmin = "^3.0.1"
django-debug-toolbar = "^4.4.6"
django-redis = "^5.4.0"
django-debug-toolbar-force = "^0.2"
django-phonenumber-field = "^8.0.0"
pyotp = "^2.9.0"
phonenumbers = "^8.13.47"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
